Moved to Concept companion object
