+++ morphir/runtime/test/jvm/src/org/finos/morphir/runtime/DefaultsSpec.scala
@@ -30,7 +30,7 @@ object DefaultsTestingSpec extends MorphirBaseSpec {
             res <- rt.evaluate(qn"Defaults:ExampleModule:expectedDefault")
               .provideEnvironment(MorphirEnv.live)
               .toZIOWith(RTExecutionContext.typeChecked)
-            default <- ZIO.fromEither(MDMDefaults.default(res.shape))
+            default <- ZIO.fromEither(res.shape.defaultData())
           } yield assertTrue(res == default)
         }
       },
@@ -40,25 +40,25 @@ object DefaultsTestingSpec extends MorphirBaseSpec {
             res <- rt.evaluate(qn"Defaults:ExampleModule:nonDefault")
               .provideEnvironment(MorphirEnv.live)
               .toZIOWith(RTExecutionContext.typeChecked)
-            default <- ZIO.fromEither(MDMDefaults.default(res.shape))
+            default <- ZIO.fromEither(res.shape.defaultData())
           } yield assertTrue(res != default)
         }
-      },
-      test("Record can be filled with defaults") {
-        ZIO.serviceWithZIO[TypedMorphirRuntime] { rt =>
-          for {
-            small <- rt.evaluate(qn"Defaults:ExampleModule:smallExample")
-              .provideEnvironment(MorphirEnv.live)
-              .toZIOWith(RTExecutionContext.typeChecked)
-            larger <- rt.evaluate(qn"Defaults:ExampleModule:largerExample")
-              .provideEnvironment(MorphirEnv.live)
-              .toZIOWith(RTExecutionContext.typeChecked)
-            largerConcept = larger.shape
-            filled  <- ZIO.fromEither(MDMDefaults.fillWithDefaults(small, largerConcept))
-            default <- ZIO.fromEither(MDMDefaults.default(largerConcept))
-          } yield assertTrue(filled == larger)
-        }
       }
+      // test("Record can be filled with defaults") {
+      //   ZIO.serviceWithZIO[TypedMorphirRuntime] { rt =>
+      //     for {
+      //       small <- rt.evaluate(qn"Defaults:ExampleModule:smallExample")
+      //         .provideEnvironment(MorphirEnv.live)
+      //         .toZIOWith(RTExecutionContext.typeChecked)
+      //       larger <- rt.evaluate(qn"Defaults:ExampleModule:largerExample")
+      //         .provideEnvironment(MorphirEnv.live)
+      //         .toZIOWith(RTExecutionContext.typeChecked)
+      //       largerConcept = larger.shape
+      //       filled  <- ZIO.fromEither(Concept.Defaults.fillWithDefaults(small, largerConcept))
+      //       default <- ZIO.fromEither(largerConcept.defaultData)
+      //     } yield assertTrue(filled == larger)
+      //   }
+      // }
     )
   ).provideLayerShared(testLayer)
 }
